Sensor compatibility + keep eyes on for a period
Sensors sometimes only generate a very short signal that would even be shorter than a debounce time for a press on abutton. So adjusted the debounce time for the trigger to 1 ms.
Also made the eyes stay on for 5 seconds fter the trigger has been activated. Will be reset to 5 seconds each time the trigger gets activated again.